Alex Hormozi Ai/backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
openai==1.3.8
langchain==0.0.335
langchain-openai==0.0.2
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# Data Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
nltk==3.8.1

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Data Processing and Analysis
pandas==2.1.3
numpy==1.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0

